Update README and error message
If Killing Floor 2 installation resides in a system partition ( e.g.
C:\Program Files (x86) ), and the script is run without administrative
privileges, it won't be able to see KFGame.exe and will print a misleading
message as a result. The error message for missing KFGame.exe and the corresponding README section have been updated.